pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-dance>=7.1.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "flask-login>=0.6.3",
    "oauthlib>=3.3.1",
    "pyjwt>=2.10.1",
    "flask-wtf>=1.2.2",
    "flask-migrate>=4.1.0",
    "wtforms>=3.2.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
    "pytest>=8.4.1",
    "requests>=2.32.4",
    "sendgrid>=6.12.4",
    "weasyprint>=65.1",
    "scikit-learn>=1.7.0",
    "pandas>=2.3.0",
    "numpy>=2.3.1",
    "apscheduler>=3.11.0",
    "joblib>=1.5.1",
    "openai>=1.91.0",
    "openpyxl>=3.1.5",
    "authlib>=1.6.0",
    "markdown>=3.8.2",
    "sentry-sdk>=2.32.0",
    "prometheus-flask-exporter>=0.23.2",
    "psutil>=7.0.0",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
]
